优先使用int替换为uint，方便运算，能够准确识别产生的负数。

+++ Config.cpp
@@ -76,8 +76,8 @@ bool ConfigBlock::Init(const String& name, const Buffer& bs)
 {
     if(!name) return false;
 
-	assert(name.Length() < sizeof(Name), "配置区名称太长");
-    if(name.Length() >= sizeof(Name)) return false;
+	assert(name.Length() < (int)sizeof(Name), "配置区名称太长");
+    if(name.Length() >= (int)sizeof(Name)) return false;
 
 	//TS("ConfigBlock::Init");
 
@@ -181,8 +181,8 @@ const ConfigBlock* FindBlock(const Storage& st, uint addr, const String& name)
 
     if(!name) return nullptr;
 
-	assert(name.Length() < sizeof(ConfigBlock::Name), "配置区名称太长");
-    if(name.Length() >= sizeof(ConfigBlock::Name)) return nullptr;
+	assert(name.Length() < (int)sizeof(ConfigBlock::Name), "配置区名称太长");
+    if(name.Length() >= (int)sizeof(ConfigBlock::Name)) return nullptr;
 
 	if(!CheckSignature(st, addr, false)) return nullptr;
 
@@ -307,8 +307,8 @@ const void* Config::Set(const String& name, const Buffer& bs) const
 
     if(!name) return nullptr;
 
-	assert(name.Length() < sizeof(ConfigBlock::Name), "配置区名称太长");
-    if(name.Length() >= sizeof(ConfigBlock::Name)) return nullptr;
+	assert(name.Length() < (int)sizeof(ConfigBlock::Name), "配置区名称太长");
+    if(name.Length() >= (int)sizeof(ConfigBlock::Name)) return nullptr;
 
 	auto cfg = FindBlock(Device, Address, name);
 	if(!cfg) cfg	= NewBlock(Device, Address, bs.Length());
